+++ sales-app/app/src/main/java/com/peyess/salesapp/data/repository/edit_service_order/payment/EditLocalPaymentRepository.kt
@@ -9,6 +9,7 @@ import com.peyess.salesapp.data.repository.edit_service_order.payment.error.Read
 import com.peyess.salesapp.data.repository.edit_service_order.payment.error.UpdateLocalPaymentError
 import com.peyess.salesapp.typing.sale.PaymentDueDateMode
 import kotlinx.coroutines.flow.Flow
+import java.math.BigDecimal
 import java.time.ZonedDateTime
 
 typealias EditLocalPaymentInsertResponse = Either<InsertLocalPaymentError, Long>
@@ -20,7 +21,7 @@ typealias EditLocalPaymentStreamResponse = Flow<EditLocalPaymentFetchResponse>
 typealias EditLocalPaymentFetchSingleResponse = Either<ReadLocalPaymentError, LocalPaymentDocument>
 typealias EditLocalPaymentStreamSingleResponse = Flow<EditLocalPaymentFetchSingleResponse>
 
-typealias EditLocalPaymentFetchTotalResponse = Either<ReadLocalPaymentError, Double>
+typealias EditLocalPaymentFetchTotalResponse = Either<ReadLocalPaymentError, BigDecimal>
 typealias EditLocalPaymentStreamTotalResponse = Flow<EditLocalPaymentFetchTotalResponse>
 
 typealias EditLocalPaymentUpdateResponse = Either<UpdateLocalPaymentError, Unit>
@@ -42,7 +43,7 @@ interface EditLocalPaymentRepository {
     suspend fun updateMethodId(paymentId: Long, methodId: String): EditLocalPaymentUpdateResponse
     suspend fun updateMethodName(paymentId: Long, methodName: String): EditLocalPaymentUpdateResponse
     suspend fun updateMethodType(paymentId: Long, methodType: String): EditLocalPaymentUpdateResponse
-    suspend fun updateValue(paymentId: Long, value: Double): EditLocalPaymentUpdateResponse
+    suspend fun updateValue(paymentId: Long, value: BigDecimal): EditLocalPaymentUpdateResponse
     suspend fun updateInstallments(paymentId: Long, installments: Int): EditLocalPaymentUpdateResponse
     suspend fun updateDocument(paymentId: Long, document: String): EditLocalPaymentUpdateResponse
     suspend fun updateCardFlagName(paymentId: Long, cardFlagName: String): EditLocalPaymentUpdateResponse
