+++ sales-app/app/src/main/java/com/peyess/salesapp/screen/edit_service_order/payment_discount/EditPaymentDiscountScreen.kt
@@ -36,10 +36,10 @@ fun EditPaymentDiscountScreen(
     DiscountScreenUI(
         modifier = modifier,
         selectedMethod = currentDiscount.method,
-        discountValue = BigDecimal(currentDiscount.value),
+        discountValue = currentDiscount.value,
         originalPrice = originalPrice,
         pricePreview = pricePreview,
-        onChangeDiscountValue = { viewModel.onChangeDiscountValue(it.toDouble()) },
+        onChangeDiscountValue = { viewModel.onChangeDiscountValue(it) },
         onChangeDiscountMethod = viewModel::onChangeDiscountMethod,
         onDone = onDone,
     )
